requirements.txt
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
Flask==2.2.5
idna==3.10
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.6
MarkupSafe==2.1.5
python-dotenv==0.21.1
PyYAML==6.0.1
qtrade==0.6.1
requests==2.31.0
typing_extensions==4.7.1
urllib3==2.0.7
Werkzeug==2.2.3
zipp==3.15.0
